Move HttpClient coroutines into header
